Change DB migration to use URI hash rather than full URL.
